Updating initMementoable to use the newer `.on` method of binding events
The `bind` and `unbind` functions were renamed to `on` and `off` in
Backbone v0.9.0. While they still work, the current documentation
doesn't even mention them. Using the newer names, in order to prevent
confusion, especially since it's to mistake `bind` for Backbone Events
with `_.bind`.

There is no need to use `_.bind`, as we can pass-on the context when
binding the event directly.